fix #61, fix #60 map and arrow rotation (#62)
The built in Zoom controls don't work with the map rotation, I had to implement them on my own. They don't have the autohide feature at the moment. I don't know if this is a problem. I'll hide them in PiP mode only.

I've merged this code for now. Feel free to open new issues if you encounter problems with it. The compass smoothing is already in a new issue.